fix for 64b
git-svn-id: svn://svn.savannah.nongnu.org/osip/trunk/osip@651 a617929b-4c5a-41d8-b177-2eda7db2ca6c

+++ src/osipparser2/osip_port.c
@@ -1086,7 +1086,7 @@ _osip_realloc (void *ptr, size_t size, char *file, unsigned short line)
 unsigned long
 osip_hash (const char *str)
 {
-  unsigned long hash = 5381;
+  unsigned int hash = 5381;
   int c;
 
   while (c = *str++)
